requirements.txt
streamlit>=1.29.0
yt-dlp>=2024.1.0
openai-whisper>=20231117
torch>=2.2.0
torchaudio>=2.2.0
pandas>=2.0.0
numpy>=1.24.0
ffmpeg-python>=0.2.0
requests>=2.31.0